requirements.txt
streamlit
psycopg2-binary
pandas
numpy